examples/teacher-version/todoapp-backend-api-e2etest-main/requirements.txt
pytest==7.4.3
pytest-bdd==7.1.0
pytest-html==4.1.1
allure-pytest==2.13.2
setuptools>=65.5.0  # Python 3.12+ 需要，提供 distutils 兼容
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2

